changes to better work with UI

+++ TwitterTopicModeling/Migrations/TwitterContextModelSnapshot.cs
@@ -54,8 +54,8 @@ namespace TwitterTopicModeling.Migrations
                     b.Property<int?>("Reportid")
                         .HasColumnType("integer");
 
-                    b.Property<int?>("TweetId")
-                        .HasColumnType("integer");
+                    b.Property<long?>("TweetId")
+                        .HasColumnType("bigint");
 
                     b.HasKey("id");
 
@@ -68,9 +68,9 @@ namespace TwitterTopicModeling.Migrations
 
             modelBuilder.Entity("TwitterTopicModeling.Database.Models.Tweet", b =>
                 {
-                    b.Property<int>("Id")
+                    b.Property<long>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("integer")
+                        .HasColumnType("bigint")
                         .HasAnnotation("Npgsql:ValueGenerationStrategy", NpgsqlValueGenerationStrategy.IdentityByDefaultColumn);
 
                     b.Property<long>("ExternalId")
